Update readinglog_priti.html
I added 3 bar charts and a pie chart. However, in order to make changes, we must click a button at the bottom left of the screen. Can one of you please find a way to call the updatecharts() javascript function every time the user makes a change rather than using the button that I put for testing?